Face_Recognition/requirements.txt
fastapi==0.101.0
mediapipe==0.8.10.1
opencv-python==4.8.1.78
opencv-contrib-python==4.8.0.74
Pillow==10.0.0
pickle
joblib==1.2.0
sklearn==0.0.post5
facenet-pytorch==2.5.2
pandas==2.1.3
protobuf==3.19.6
